requirements.txt
AHRS==0.3.1
matplotlib==3.9.2
numpy==2.0.1
opencv-python==4.10.0.84
pyzmq==26.1.0
